refactoring created interfaces in interface folder
