+++ Rdc-Calculator/Eurozone-Calculator/interfaces/IEuropeanHumanRightsCourt.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Eurozone_Calculator.interfaces
+{
+        interface IEuropeanHumanRightsCourt
+    {
+        void IsStateRespectingHumanRights();
+    }
+}
