+++ Rdc-Calculator/Eurozone-Calculator/interfaces/IEuropeanStockMarket.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Eurozone_Calculator.interfaces
+{
+    interface IEuropeanStockMarket
+    {
+
+        void CalculateSpread();
+        void DisplayGovernmentBondValue();
+
+    }
+}
